flask_backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
supabase>=2.8.0
PyJWT==2.8.0
Flask-Bcrypt==1.0.1
pytest==7.4.3
firebase-admin==6.5.0
